fixed abundance table formatting & output
 - added missing taxon column for unclassified reads in abundance tables
 - prevent that read numbers are shown in scientific notation

+++ src/classification.h
@@ -54,8 +54,7 @@ struct rank_higher {
     }
 };
 
-using taxon_count_map = std::map<const taxon*, float, rank_higher>;
-// using taxon_count_map = std::unordered_map<const taxon*, query_id>;
+using taxon_count_map = std::map<const taxon*, double, rank_higher>;
 
 
 /*************************************************************************//**
